Reuse test_ops for XPU
[ghstack-poisoned]
