+++ test/test_ops.py
@@ -2407,6 +2407,9 @@ class TestFakeTensor(TestCase):
             strided_result = op(sample.input, *sample.args, **kwargs)
             self.assertEqual(strided_result.layout, torch.strided)
 
+class Namespace:
+    class TestCommonWrapper(TestCommon):
+        pass
 
 instantiate_device_type_tests(TestCommon, globals())
 instantiate_device_type_tests(TestCompositeCompliance, globals())
